+++ test/test_helper.rb
@@ -1,7 +1,9 @@
 # frozen_string_literal: true
 
 require "simplecov"
-SimpleCov.start
+SimpleCov.start do
+  add_filter(/test/)
+end
 
 require "bundler/setup"
 require "i18n-js"